+++ src/test/java/org/example/puppy/withdraw/ChangingAccountBalanceTest.java
@@ -14,7 +14,7 @@ import requstprovider.PuppyRequestProvider;
 
 public class ChangingAccountBalanceTest extends BaseWebTest {
     private static final int ID = 1;
-    private static final double ACCOUNT_BALANCE = 1;
+    private static final float ACCOUNT_BALANCE = 1;
     private final LoginPage loginPage = new LoginPage();
     private final BasePage basePage = new BasePage();
     private final NewWithdrawPage newWithdrawPage = new NewWithdrawPage();
